refactor: 모듈화 진행 + 이벤트 핸들러 사용
