+++ src/script/error/modules/addErrorMessage.js
@@ -1,0 +1,5 @@
+// 에러 메세지 추가
+export default function addErrorMessage(value, str) {
+    value.innerText = str;
+    value.style.display = "block";
+}
